requirements.txt
anthropic>=0.25.0
sqlglot>=20.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
